:recycle: Re-authenticate only with custom auth token
It's important to understand that AccountManager is not an encryption
service or a keychain. It stores account credentials just as you pass
them, in plain text. On most devices, this isn't a particular concern,
because it stores them in a database that is only accessible to root.
But on a rooted device, the credentials would be readable by anyone with
adb access to the device.

With this in mind, the account password will no longer be stored in the
account manager. The re-authentication will happend using the custom
auth token obtained from AccountManager.peekToken.

Given a token is obtained from the cache, for the sake of simplicity:

  - The expiration datetime will not be checked.
  - The token will be cached indefinitely until the app explicitely
    invidates it (i.e. customTokens=false).
  - If the sign in fails, the user will be prompted to the authenticator
    activity.

References:

  - Google Firebase Authentication
    See https://firebase.google.com/docs/auth/android/manage-users

  - Google Android Documentation - KEY_CUSTOM_TOKEN_EXPIRY
    See https://goo.gl/Xtk053

  - AccountManager.peekAuthToken returning null value.
    See http://www.stackoverflow.com/questions/29657289

  - Do Firebase tokens expire?
    See http://www.stackoverflow.com/questions/39091833/

  - JSON Web Tokens documentation
    See https://jwt.io/introduction/

  - RS256 vs HS256: What's the difference?
    See http://stackoverflow.com/questions/39239051

  - Google Training - Custom Auth: Be Smart about Security!
    https://developer.android.com/training/id-auth/custom_auth.html